fixed authcontext bug
